Add option to switch tmux statusline per instance
Added definable option to integration option set to allow tmux to be set
to only change its current instance's statusline instead of all tmux
instances globally.

By default this is turned off (to mimick the existing behavior) but can
be turned on by changing the configuration to:

require("true-zen").setup({
    integrations = {
        tmux = { global = false },
    }
})

which changes TrueZen to afterwards only affect the tmux instance
containing the neovim instance calling it.